requirements.txt
Django==1.9.13
django-filter==1.1.0
django-tastypie==0.14.0
djangorestframework==3.7.1
Markdown==2.6.9
mysqlclient==1.3.12
olefile==0.44
Pillow==4.2.1
pygame==1.9.3
python-dateutil==2.6.1
python-mimeparse==1.6.0
reportlab==3.4.0
six==1.11.0
numpy==1.13.3
scipy==1.0.0